+++ src/com/aryan/rain/entity/Entity.java
@@ -9,7 +9,7 @@ import java.util.Random;
 
 // A Entity doesn't necessarily have sprites.
 public class Entity {
-    protected int x, y;
+    protected double x, y;
     protected Sprite sprite;
     private boolean removed = false;
     protected Level level;
@@ -28,7 +28,7 @@ public class Entity {
     }
 
     public void render(Screen screen){
-        if (sprite != null) screen.renderSprite(x,y, sprite, true);
+        if (sprite != null) screen.renderSprite((int)x, (int)y, sprite, true);
     }
 
     // Remove entity from level
@@ -36,11 +36,11 @@ public class Entity {
         removed = true;
     }
 
-    public int getX(){
+    public double getX(){
         return x;
     }
 
-    public int getY(){
+    public double getY(){
         return y;
     }
